resolve issue in login screen , move AsteriskVisualTransformation
