+++ presentation/ui/src/main/res/drawable/eye.xml
